Add RSA Encryption and Decryption, Add Byte Hex String conversion
Use Java's KeyPairGenerator with RSA algorithm for generating public
and private key.

Use Java's Cipher for encryption and decryption.

Implement a simple byte array to hex string converter, and vice versa.

App now successfully encrypts the contents of edittext, displays in
textview, and then decrypt button converts back again.

The app crashes when string becomes to long i.e. longer than the 1024
keysize. RSA not ideal for encrypting large amounts of data, so an
improvement would be to encrypt the string with standard session key
with symmetric algorithm such as AES, then encrypt the session key
with RSA algorithm, and store the encrypted data and the encrypted AES
key.